refactor : 상품 상세 정보 클래스명 변경
